+++ src/ClashRoyale/Logic/Sessions/Session.cs
@@ -8,7 +8,7 @@ namespace ClashRoyale.Logic.Sessions
         [JsonProperty("location")] public Location Location = new Location();
         [JsonIgnore] public DateTime SessionStart = DateTime.UtcNow;
         [JsonProperty("ip")] public string Ip { get; set; }
-        [JsonProperty("duration")] public long Duration { get; set; }
+        [JsonProperty("duration")] public int Duration { get; set; }
         [JsonProperty("startDate")] public string StartDate { get; set; }
         [JsonProperty("deviceCode")] public string DeviceCode { get; set; }
         [JsonProperty("gameVersion")] public string GameVersion { get; set; }
